pyproject.toml
[tool.poetry]
name = "assistant"
version = "0.1.0"
description = ""
authors = ["gal16v8d"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
yfinance = "^0.2.63"
selenium = "^4.33.0"
speechrecognition = "^3.14.3"
pyjokes = "^0.8.3"
pyttsx3 = "^2.98"
pyaudio = "^0.2.14"
google-genai = "^1.20.0"
pillow = "^11.2.1"

[tool.poetry.group.dev.dependencies]
pylint = "^3.3.7"
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
